requirements.txt
click==7.1.1
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
PyYAML==5.3.1
Werkzeug==1.0.1
